% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageOverlap.R
\name{average_overlap}
\alias{average_overlap}
\title{Compute the average overlap}
\usage{
average_overlap(obj)
}
\arguments{
\item{obj}{Either a vector or matrix}
}
\value{
A vector of the average overlap
}
\description{
Compute the average overlap
}

