% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Quantify overlap of two lists}
\usage{
overlap(l1, l2, depth = NULL)
}
\arguments{
\item{l1}{First list}

\item{l2}{Second list}

\item{depth}{Interest margin, i.e., agreement is limited to the first \code{depth} elements of each list.}
}
\value{
A vector of length \code{depth} whose elements are the average overlap.
}
\description{
Quantify the average overlap of two lists
in the head of the lists and until given depth
}
\examples{
overlap(c(1,2,3,4,5,7,6,8), c(1,2,3,5,6,4,7,8))
}
\author{
Claus Ekstrøm <ekstrom@sund.ku.dk>, Thomas A. Gerds <tag@biostat.ku.dk>
}
\seealso{
list.agreement
}

