% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.agreement.R
\name{plot.agreement}
\alias{plot.agreement}
\title{Plot agreement}
\usage{
\method{plot}{agreement}(x, xlim, ylim, xlab = "List depth",
  ylab = "Agreement", add = FALSE, ...)
}
\arguments{
\item{x}{Agreement object}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{xlab}{x-axis lab}

\item{ylab}{y-axis lab}

\item{...}{Processed by \code{\link{prodlim::SmartControl}}.}
}
\value{
Graph
}
\description{
Plot the agreement between lists as a function of the list depth
}
\examples{
a <- agreement(list(c(1,2,3,4,5,7,6,8,9,10,11,12,13), c(5,11,4,7,8,3,12,13,6,10,9,2,1)))
plot(a)
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}

