% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sra.R
\name{sra}
\alias{sra}
\title{Compute the sequential rank agreement}
\usage{
sra(object, B = 1)
}
\arguments{
\item{B}{An integer giving the number of randomization to sample over in the case of censored observations}

\item{obj}{Either matrix where each column is a ranked list of items or a list of ranked lists of items. Elements are integers between 1 and the length of the lists. The lists should have the same length but censoring can be used by setting the list to zero from a point onwards. See details for more information.}
}
\value{
A vector of the sequential rank agreement
}
\description{
Compute the sequential rank agreement
}
\author{
Claus Ekstrøm <ekstrom@sund.ku.dk>
}

