% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperRanker-package.R
\docType{package}
\name{SuperRanker-package}
\alias{SuperRanker}
\alias{SuperRanker-package}
\title{SuperRanker: Sequential Rank Agreement}
\description{
Tools for analysing the agreement of two or more rankings of the same items. Examples are importance rankings of predictor variables and risk predictions of subjects. Benchmarks for agreement are computed based on random permutation and bootstrap. See Ekstrøm CT, Gerds TA, Jensen, AK (2018). "Sequential rank agreement methods for comparison of ranked lists." _Biostatistics_, *20*(4), 582-598 \doi{10.1093/biostatistics/kxy017} for more information.
}
\author{
\strong{Maintainer}: Claus Thorn Ekstrøm \email{ekstrom@sund.ku.dk}

Authors:
\itemize{
  \item Thomas Alexander Gerds \email{tag@biostat.ku.dk}
}

}
\keyword{internal}
