% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SuperRanker-package.R, R/SuperRanker.R
\docType{package}
\name{SuperRanker}
\alias{SuperRanker}
\alias{SuperRanker-package}
\title{Functions related to comparison of ranked lists}
\description{
SuperRanker allows you tRoxygen2 allows you to write documentation in comment blocks co-located
with code.

SuperRanker allows you tRoxygen2 allows you to write documentation in comment blocks co-located
with code.
}
\details{
The only function you're likely to need from \pkg{roxygen2} is
\code{\link{roxygenize}}. Otherwise refer to the vignettes to see
how to format the documentation.

The only function you're likely to need from \pkg{roxygen2} is
\code{\link{roxygenize}}. Otherwise refer to the vignettes to see
how to format the documentation.
}

