% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageOverlap.R
\name{average_overlap}
\alias{average_overlap}
\title{Compute the average overlap}
\usage{
average_overlap(obj)
}
\arguments{
\item{obj}{Either a vector or matrix}
}
\value{
A vector of the average overlap
}
\description{
Compute the average overlap
}
\examples{
# setting with 3 lists
mlist <- matrix(cbind(1:8,c(1,2,3,5,6,7,4,8),c(1,5,3,4,2,8,7,6)),ncol=3)
average_overlap(mlist)
}

