% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R, R/overlap.R
\name{overlap}
\alias{overlap}
\title{Compute the overlap between k ranked lists}
\usage{
overlap(l1, l2, depth = NULL)

overlap(l1, l2, depth = NULL)
}
\arguments{
\item{l1}{First list}

\item{l2}{Second list}

\item{depth}{Interest margin, i.e., agreement is limited to the first \code{depth} elements of each list.}

\item{rankMat}{A matrix with k columns corresponding to the k ranked lists. Elements of each column are integers between 1 and the length of the lists}
}
\value{
A vector of the same length as the rows in rankMat containing the overlap between the lists for each rank

A vector of length \code{depth} whose elements are the average overlap.
}
\description{
Computes the overlap (number of items present in all k lists divided by the current rank) for each rank in the k lists

Quantify the average overlap of two lists
in the head of the lists and until given depth
}
\examples{
overlap(c(1,2,3,4,5,7,6,8), c(1,2,3,5,6,4,7,8))
}
\author{
Claus Ekstrøm <ekstrom@sund.ku.dk>

Claus Ekstrøm <ekstrom@sund.ku.dk>, Thomas A. Gerds <tag@biostat.ku.dk>
}
\seealso{
list.agreement
}

