% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overlap}
\alias{overlap}
\title{Compute the overlap between k ranked lists}
\usage{
overlap(rankMat)
}
\arguments{
\item{rankMat}{A matrix with k columns corresponding to the k ranked lists. Elements of each column are integers between 1 and the length of the lists}
}
\value{
A vector of the same length as the rows in rankMat containing the overlap between the lists for each rank
}
\description{
Computes the overlap (number of items present in all k lists divided by the current rank) for each rank in the k lists
}
\author{
Claus Ekstrøm <ekstrom@sund.ku.dk>
}
