\name{random_list_sra}
\alias{random_list_sra}
\title{Simulate sequential rank agreement for randomized unrelated lists}
\usage{
random_list_sra(object, B = 1, n = 1)
}
\arguments{
  \item{obj}{Either a vector or matrix}

  \item{B}{Either a vector or matrix}

  \item{n}{the number of sequential rank agreement curves
  to produce}
}
\value{
A matrix with n columns each representing the sequential
rank agreement obtained from
}
\description{
Simulate sequential rank agreement for randomized unrelated
lists
}
\author{
Claus Ekstrøm <ekstrom@sund.ku.dk>
}

