% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sra.R
\name{smooth_sra}
\alias{smooth_sra}
\title{Smooth quantiles of a matrix of sequential ranked agreements}
\usage{
smooth_sra(object, confidence = 0.95)
}
\arguments{
\item{confidence}{the limits to compute}

\item{obj}{A matrix}
}
\value{
A list containing two vectors for the smoothed lower and upper limits
}
\description{
Smooth quantiles of a matrix of sequential ranked agreements
}
\author{
Claus Ekstrøm <ekstrom@sund.ku.dk>
}

