% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sracppfull}
\alias{sracppfull}
\title{Compute the sequential rank agreement between k ranked lists}
\usage{
sracppfull(rankMat, type = 0L)
}
\arguments{
\item{rankMat}{A matrix with k columns corresponding to the k ranked lists. Elements of each column are integers between 1 and the length of the lists}

\item{type}{The type of distance measure to use: 0 (the default) is the variance while 1 is MAD (mean absolute deviation)}
}
\value{
A vector of the same length as the rows in rankMat containing the sequential rank agreement between the lists for each depth (squared for type=0)
}
\description{
Computes the sequential rank agreement (number of items present in all k lists divided by the current rank) for each rank in the k lists
}
\author{
Claus Ekstrøm <ekstrom@sund.ku.dk>
}

