% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperRanker-package.R
\docType{package}
\name{SuperRanker}
\alias{SuperRanker}
\alias{SuperRanker-package}
\title{Functions related to comparison of ranked lists}
\description{
SuperRanker allows you to estimate the agreement between two or more rankings of the same items.
}

